# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: arc — Personal Brand Strategy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/schedule-batch:
    post:
      operationId: scheduleBatchPosts
      tags: [posts]
      summary: Schedule a batch of posts across dates
      description: Assigns target dates to several posts at once, spreading them from a start date by a fixed day interval, and marks each as scheduled.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleBatchInput"
      responses:
        "200":
          description: The updated posts after scheduling
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "400":
          description: Invalid input
        "404":
          description: No client profile yet
  /posts/draft:
    post:
      operationId: draftPosts
      tags: [posts]
      summary: Draft platform-appropriate copy in the client's voice (Ghostwriter)
      description: Generates one or more editable content drafts (post, hook, or article) grounded in the client's narrative, voice/tone, and profile. Drafts are returned for review and are NOT persisted; the client edits and saves the ones they want via the normal create-post route.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftPostsInput"
      responses:
        "200":
          description: Generated drafts for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftPostsResult"
        "400":
          description: Invalid input
        "404":
          description: No client profile yet
        "502":
          description: Generation failed
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /posts/{id}/handoff:
    post:
      operationId: handoffPost
      tags: [posts]
      summary: Push a single post into the client's connected scheduler
      description: Sends the post's content (and schedule date, if any) to the client's own connected third-party scheduler. arc never publishes directly. Records the hand-off state on the post.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffInput"
      responses:
        "200":
          description: The post after a successful hand-off
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid input
        "404":
          description: Post not found
        "409":
          description: No scheduler connected for that provider
        "502":
          description: The scheduler rejected the post
  /posts/handoff-batch:
    post:
      operationId: handoffBatchPosts
      tags: [posts]
      summary: Push several posts into the client's connected scheduler
      description: Hands off a batch of posts at once. Returns a per-post result so partial failures are visible, plus the updated posts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandoffBatchInput"
      responses:
        "200":
          description: Per-post hand-off results and the updated posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandoffBatchResult"
        "400":
          description: Invalid input
        "404":
          description: No client profile yet
        "409":
          description: No scheduler connected for that provider
...
components:
  schemas:
...
    Post:
      type: object
      required: [id, title, content, platform, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        platform:
          type: string
          enum: [linkedin, twitter, instagram, blog, other]
        status:
          type: string
          enum: [draft, scheduled, published]
        scheduledAt:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        handoffProvider:
          type: ["string", "null"]
        handoffAt:
          type: ["string", "null"]
        handoffRef:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    HandoffInput:
      type: object
      properties:
        provider:
          type: string
          description: Which connected scheduler to push to. Defaults to the only connected provider when omitted.
        scheduledAt:
          type: ["string", "null"]
          description: ISO 8601 schedule date/time. When omitted, the post's own scheduledAt (or the scheduler's next free slot) is used.
    HandoffBatchInput:
      type: object
      required: [postIds]
      properties:
        postIds:
          type: array
          items:
            type: integer
        provider:
          type: string
    HandoffResultItem:
      type: object
      required: [postId, ok]
      properties:
        postId:
          type: integer
        ok:
          type: boolean
        error:
          type: ["string", "null"]
    HandoffBatchResult:
      type: object
      required: [results, posts]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/HandoffResultItem"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
...
    PostInput:
      type: object
      required: [title, content, platform, status]
      properties:
        title:
          type: string
        content:
          type: string
        platform:
          type: string
          enum: [linkedin, twitter, instagram, blog, other]
        status:
          type: string
          enum: [draft, scheduled, published]
        scheduledAt:
          type: string
        tags:
          type: array
          items:
            type: string
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        platform:
          type: string
          enum: [linkedin, twitter, instagram, blog, other]
        status:
          type: string
          enum: [draft, scheduled, published]
        scheduledAt:
          type: string
        tags:
          type: array
          items:
            type: string
    DraftPostsInput:
      type: object
      required: [format, platform]
      properties:
        format:
          type: string
          enum: [post, hook, article]
          description: The kind of copy to draft — a full platform post, short scroll-stopping hooks, or a longer-form article.
        platform:
          type: string
          enum: [linkedin, twitter, instagram, blog, other]
        brief:
          type: string
          description: Optional freeform topic, angle, or instruction the draft should be about.
        ideaId:
          type: integer
          description: Optional saved idea to ground the draft in (its title and notes are used).
        theme:
          type: string
          description: Optional narrative theme or angle title to anchor the draft to.
        count:
          type: integer
          description: How many draft variants to produce (server-bounded).
        feedback:
          type: string
          description: Optional revision notes to steer a regeneration.
    DraftedPost:
      type: object
      required: [title, content, format]
      properties:
        title:
          type: string
        content:
          type: string
        format:
          type: string
          enum: [post, hook, article]
    DraftPostsResult:
      type: object
      required: [drafts]
      properties:
        drafts:
          type: array
          items:
            $ref: "#/components/schemas/DraftedPost"
...
    ScheduleBatchInput:
      type: object
      required: [postIds, startDate]
      properties:
        postIds:
          type: array
          items:
            type: integer
          description: Post IDs to schedule, in the order they should be laid out.
        startDate:
          type: string
          description: ISO date for the first post (YYYY-MM-DD).
        intervalDays:
          type: integer
          description: Days between consecutive posts. Defaults to 1.
        time:
          type: string
          description: Time of day in HH:MM (24h). Defaults to 09:00.
...